--- modules/apps/batch-planner/batch-planner-rest-api/src/main/java/com/liferay/headless/admin/batch/planner/dto/v1_0/Log.java
@@ -1,496 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.admin.batch.planner.dto.v1_0;
-
-import com.fasterxml.jackson.annotation.JsonFilter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import com.liferay.petra.function.UnsafeSupplier;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLField;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLName;
-import com.liferay.portal.vulcan.util.ObjectMapperUtil;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-import java.io.Serializable;
-
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-
-import javax.annotation.Generated;
-
-import javax.xml.bind.annotation.XmlRootElement;
-
-/**
- * @author Matija Petanjek
- * @generated
- */
-@Generated("")
-@GraphQLName("Log")
-@JsonFilter("Liferay.Vulcan")
-@XmlRootElement(name = "Log")
-public class Log implements Serializable {
-
-	public static Log toDTO(String json) {
-		return ObjectMapperUtil.readValue(Log.class, json);
-	}
-
-	@Schema
-	public String getDispatchTriggerExternalReferenceCode() {
-		return dispatchTriggerExternalReferenceCode;
-	}
-
-	public void setDispatchTriggerExternalReferenceCode(
-		String dispatchTriggerExternalReferenceCode) {
-
-		this.dispatchTriggerExternalReferenceCode =
-			dispatchTriggerExternalReferenceCode;
-	}
-
-	@JsonIgnore
-	public void setDispatchTriggerExternalReferenceCode(
-		UnsafeSupplier<String, Exception>
-			dispatchTriggerExternalReferenceCodeUnsafeSupplier) {
-
-		try {
-			dispatchTriggerExternalReferenceCode =
-				dispatchTriggerExternalReferenceCodeUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String dispatchTriggerExternalReferenceCode;
-
-	@Schema
-	public String getExportTaskExternalReferenceCode() {
-		return exportTaskExternalReferenceCode;
-	}
-
-	public void setExportTaskExternalReferenceCode(
-		String exportTaskExternalReferenceCode) {
-
-		this.exportTaskExternalReferenceCode = exportTaskExternalReferenceCode;
-	}
-
-	@JsonIgnore
-	public void setExportTaskExternalReferenceCode(
-		UnsafeSupplier<String, Exception>
-			exportTaskExternalReferenceCodeUnsafeSupplier) {
-
-		try {
-			exportTaskExternalReferenceCode =
-				exportTaskExternalReferenceCodeUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String exportTaskExternalReferenceCode;
-
-	@Schema
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	@JsonIgnore
-	public void setId(UnsafeSupplier<Long, Exception> idUnsafeSupplier) {
-		try {
-			id = idUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Long id;
-
-	@Schema
-	public String getImportTaskExternalReferenceCode() {
-		return importTaskExternalReferenceCode;
-	}
-
-	public void setImportTaskExternalReferenceCode(
-		String importTaskExternalReferenceCode) {
-
-		this.importTaskExternalReferenceCode = importTaskExternalReferenceCode;
-	}
-
-	@JsonIgnore
-	public void setImportTaskExternalReferenceCode(
-		UnsafeSupplier<String, Exception>
-			importTaskExternalReferenceCodeUnsafeSupplier) {
-
-		try {
-			importTaskExternalReferenceCode =
-				importTaskExternalReferenceCodeUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String importTaskExternalReferenceCode;
-
-	@Schema
-	public Long getPlanId() {
-		return planId;
-	}
-
-	public void setPlanId(Long planId) {
-		this.planId = planId;
-	}
-
-	@JsonIgnore
-	public void setPlanId(
-		UnsafeSupplier<Long, Exception> planIdUnsafeSupplier) {
-
-		try {
-			planId = planIdUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Long planId;
-
-	@Schema
-	public Integer getSize() {
-		return size;
-	}
-
-	public void setSize(Integer size) {
-		this.size = size;
-	}
-
-	@JsonIgnore
-	public void setSize(UnsafeSupplier<Integer, Exception> sizeUnsafeSupplier) {
-		try {
-			size = sizeUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Integer size;
-
-	@Schema
-	public Integer getStatus() {
-		return status;
-	}
-
-	public void setStatus(Integer status) {
-		this.status = status;
-	}
-
-	@JsonIgnore
-	public void setStatus(
-		UnsafeSupplier<Integer, Exception> statusUnsafeSupplier) {
-
-		try {
-			status = statusUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Integer status;
-
-	@Schema
-	public Integer getTotal() {
-		return total;
-	}
-
-	public void setTotal(Integer total) {
-		this.total = total;
-	}
-
-	@JsonIgnore
-	public void setTotal(
-		UnsafeSupplier<Integer, Exception> totalUnsafeSupplier) {
-
-		try {
-			total = totalUnsafeSupplier.get();
-		}
-		catch (RuntimeException re) {
-			throw re;
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Integer total;
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof Log)) {
-			return false;
-		}
-
-		Log log = (Log)object;
-
-		return Objects.equals(toString(), log.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		StringBundler sb = new StringBundler();
-
-		sb.append("{");
-
-		if (dispatchTriggerExternalReferenceCode != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"dispatchTriggerExternalReferenceCode\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(dispatchTriggerExternalReferenceCode));
-
-			sb.append("\"");
-		}
-
-		if (exportTaskExternalReferenceCode != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"exportTaskExternalReferenceCode\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(exportTaskExternalReferenceCode));
-
-			sb.append("\"");
-		}
-
-		if (id != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"id\": ");
-
-			sb.append(id);
-		}
-
-		if (importTaskExternalReferenceCode != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"importTaskExternalReferenceCode\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(importTaskExternalReferenceCode));
-
-			sb.append("\"");
-		}
-
-		if (planId != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"planId\": ");
-
-			sb.append(planId);
-		}
-
-		if (size != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"size\": ");
-
-			sb.append(size);
-		}
-
-		if (status != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"status\": ");
-
-			sb.append(status);
-		}
-
-		if (total != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"total\": ");
-
-			sb.append(total);
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	@Schema(
-		accessMode = Schema.AccessMode.READ_ONLY,
-		defaultValue = "com.liferay.headless.admin.batch.planner.dto.v1_0.Log",
-		name = "x-class-name"
-	)
-	public String xClassName;
-
-	private static String _escape(Object object) {
-		String string = String.valueOf(object);
-
-		return string.replaceAll("\"", "\\\\\"");
-	}
-
-	private static boolean _isArray(Object value) {
-		if (value == null) {
-			return false;
-		}
-
-		Class<?> clazz = value.getClass();
-
-		return clazz.isArray();
-	}
-
-	private static String _toJSON(Map<String, ?> map) {
-		StringBuilder sb = new StringBuilder("{");
-
-		@SuppressWarnings("unchecked")
-		Set set = map.entrySet();
-
-		@SuppressWarnings("unchecked")
-		Iterator<Map.Entry<String, ?>> iterator = set.iterator();
-
-		while (iterator.hasNext()) {
-			Map.Entry<String, ?> entry = iterator.next();
-
-			sb.append("\"");
-			sb.append(entry.getKey());
-			sb.append("\": ");
-
-			Object value = entry.getValue();
-
-			if (_isArray(value)) {
-				sb.append("[");
-
-				Object[] valueArray = (Object[])value;
-
-				for (int i = 0; i < valueArray.length; i++) {
-					if (valueArray[i] instanceof String) {
-						sb.append("\"");
-						sb.append(valueArray[i]);
-						sb.append("\"");
-					}
-					else {
-						sb.append(valueArray[i]);
-					}
-
-					if ((i + 1) < valueArray.length) {
-						sb.append(", ");
-					}
-				}
-
-				sb.append("]");
-			}
-			else if (value instanceof Map) {
-				sb.append(_toJSON((Map<String, ?>)value));
-			}
-			else if (value instanceof String) {
-				sb.append("\"");
-				sb.append(value);
-				sb.append("\"");
-			}
-			else {
-				sb.append(value);
-			}
-
-			if (iterator.hasNext()) {
-				sb.append(", ");
-			}
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-}
