--- modules/apps/batch-planner/batch-planner-rest-api/src/main/java/com/liferay/headless/admin/batch/planner/resource/v1_0/PlanResource.java
@@ -1,149 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.admin.batch.planner.resource.v1_0;
-
-import com.liferay.headless.admin.batch.planner.dto.v1_0.Plan;
-import com.liferay.portal.kernel.search.filter.Filter;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.ResourceActionLocalService;
-import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.odata.filter.ExpressionConvert;
-import com.liferay.portal.odata.filter.FilterParserProvider;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.pagination.Pagination;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-
-import javax.annotation.Generated;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import javax.ws.rs.core.UriInfo;
-
-import org.osgi.annotation.versioning.ProviderType;
-
-/**
- * To access this resource, run:
- *
- *     curl -u your@email.com:yourpassword -D - http://localhost:8080/o/headless-admin-batch-planner/v1.0
- *
- * @author Matija Petanjek
- * @generated
- */
-@Generated("")
-@ProviderType
-public interface PlanResource {
-
-	public static Builder builder() {
-		return FactoryHolder.factory.create();
-	}
-
-	public Page<Plan> getPlansPage(Pagination pagination) throws Exception;
-
-	public Plan postPlan(Plan plan) throws Exception;
-
-	public void deletePlan(Long planId) throws Exception;
-
-	public Plan getPlan(Long planId) throws Exception;
-
-	public Plan patchPlan(Long planId, Plan plan) throws Exception;
-
-	public default void setContextAcceptLanguage(
-		AcceptLanguage contextAcceptLanguage) {
-	}
-
-	public void setContextCompany(
-		com.liferay.portal.kernel.model.Company contextCompany);
-
-	public default void setContextHttpServletRequest(
-		HttpServletRequest contextHttpServletRequest) {
-	}
-
-	public default void setContextHttpServletResponse(
-		HttpServletResponse contextHttpServletResponse) {
-	}
-
-	public default void setContextUriInfo(UriInfo contextUriInfo) {
-	}
-
-	public void setContextUser(
-		com.liferay.portal.kernel.model.User contextUser);
-
-	public void setExpressionConvert(
-		ExpressionConvert<Filter> expressionConvert);
-
-	public void setFilterParserProvider(
-		FilterParserProvider filterParserProvider);
-
-	public void setGroupLocalService(GroupLocalService groupLocalService);
-
-	public void setResourceActionLocalService(
-		ResourceActionLocalService resourceActionLocalService);
-
-	public void setResourcePermissionLocalService(
-		ResourcePermissionLocalService resourcePermissionLocalService);
-
-	public void setRoleLocalService(RoleLocalService roleLocalService);
-
-	public default Filter toFilter(String filterString) {
-		return toFilter(
-			filterString, Collections.<String, List<String>>emptyMap());
-	}
-
-	public default Filter toFilter(
-		String filterString, Map<String, List<String>> multivaluedMap) {
-
-		return null;
-	}
-
-	public static class FactoryHolder {
-
-		public static volatile Factory factory;
-
-	}
-
-	@ProviderType
-	public interface Builder {
-
-		public PlanResource build();
-
-		public Builder checkPermissions(boolean checkPermissions);
-
-		public Builder httpServletRequest(
-			HttpServletRequest httpServletRequest);
-
-		public Builder httpServletResponse(
-			HttpServletResponse httpServletResponse);
-
-		public Builder preferredLocale(Locale preferredLocale);
-
-		public Builder user(com.liferay.portal.kernel.model.User user);
-
-	}
-
-	@ProviderType
-	public interface Factory {
-
-		public Builder create();
-
-	}
-
-}
