+++ modules/apps/batch-planner/batch-planner-rest-api/src/main/resources/com/liferay/batch/planner/rest/dto/v1_0/packageinfo
