+++ modules/apps/batch-planner/batch-planner-rest-client/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Liferay Headless Admin Batch Planner Client
-Bundle-SymbolicName: com.liferay.headless.admin.batch.planner.client
+Bundle-Name: Liferay Batch Planner REST Client
+Bundle-SymbolicName: com.liferay.batch.planner.rest.client
 Bundle-Version: 1.0.0
