--- modules/apps/batch-planner/batch-planner-rest-client/src/main/java/com/liferay/headless/admin/batch/planner/client/aggregation/Aggregation.java
@@ -1,39 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.admin.batch.planner.client.aggregation;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.annotation.Generated;
-
-/**
- * @author Matija Petanjek
- * @generated
- */
-@Generated("")
-public class Aggregation {
-
-	public Map<String, String> getAggregationTerms() {
-		return _aggregationTerms;
-	}
-
-	public void setAggregationTerms(Map<String, String> aggregationTerms) {
-		_aggregationTerms = aggregationTerms;
-	}
-
-	private Map<String, String> _aggregationTerms = new HashMap<>();
-
-}
