--- modules/apps/batch-planner/batch-planner-rest-client/src/main/java/com/liferay/headless/admin/batch/planner/client/dto/v1_0/Plan.java
@@ -1,253 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.admin.batch.planner.client.dto.v1_0;
-
-import com.liferay.headless.admin.batch.planner.client.function.UnsafeSupplier;
-import com.liferay.headless.admin.batch.planner.client.serdes.v1_0.PlanSerDes;
-
-import java.io.Serializable;
-
-import java.util.Objects;
-
-import javax.annotation.Generated;
-
-/**
- * @author Matija Petanjek
- * @generated
- */
-@Generated("")
-public class Plan implements Cloneable, Serializable {
-
-	public static Plan toDTO(String json) {
-		return PlanSerDes.toDTO(json);
-	}
-
-	public Boolean getActive() {
-		return active;
-	}
-
-	public void setActive(Boolean active) {
-		this.active = active;
-	}
-
-	public void setActive(
-		UnsafeSupplier<Boolean, Exception> activeUnsafeSupplier) {
-
-		try {
-			active = activeUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Boolean active;
-
-	public Boolean getExport() {
-		return export;
-	}
-
-	public void setExport(Boolean export) {
-		this.export = export;
-	}
-
-	public void setExport(
-		UnsafeSupplier<Boolean, Exception> exportUnsafeSupplier) {
-
-		try {
-			export = exportUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Boolean export;
-
-	public String getExternalType() {
-		return externalType;
-	}
-
-	public void setExternalType(String externalType) {
-		this.externalType = externalType;
-	}
-
-	public void setExternalType(
-		UnsafeSupplier<String, Exception> externalTypeUnsafeSupplier) {
-
-		try {
-			externalType = externalTypeUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String externalType;
-
-	public String getExternalURL() {
-		return externalURL;
-	}
-
-	public void setExternalURL(String externalURL) {
-		this.externalURL = externalURL;
-	}
-
-	public void setExternalURL(
-		UnsafeSupplier<String, Exception> externalURLUnsafeSupplier) {
-
-		try {
-			externalURL = externalURLUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String externalURL;
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public void setId(UnsafeSupplier<Long, Exception> idUnsafeSupplier) {
-		try {
-			id = idUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Long id;
-
-	public String getInternalClassName() {
-		return internalClassName;
-	}
-
-	public void setInternalClassName(String internalClassName) {
-		this.internalClassName = internalClassName;
-	}
-
-	public void setInternalClassName(
-		UnsafeSupplier<String, Exception> internalClassNameUnsafeSupplier) {
-
-		try {
-			internalClassName = internalClassNameUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String internalClassName;
-
-	public Mapping[] getMappings() {
-		return mappings;
-	}
-
-	public void setMappings(Mapping[] mappings) {
-		this.mappings = mappings;
-	}
-
-	public void setMappings(
-		UnsafeSupplier<Mapping[], Exception> mappingsUnsafeSupplier) {
-
-		try {
-			mappings = mappingsUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Mapping[] mappings;
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public void setName(UnsafeSupplier<String, Exception> nameUnsafeSupplier) {
-		try {
-			name = nameUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected String name;
-
-	public Policy[] getPolicies() {
-		return policies;
-	}
-
-	public void setPolicies(Policy[] policies) {
-		this.policies = policies;
-	}
-
-	public void setPolicies(
-		UnsafeSupplier<Policy[], Exception> policiesUnsafeSupplier) {
-
-		try {
-			policies = policiesUnsafeSupplier.get();
-		}
-		catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	protected Policy[] policies;
-
-	@Override
-	public Plan clone() throws CloneNotSupportedException {
-		return (Plan)super.clone();
-	}
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof Plan)) {
-			return false;
-		}
-
-		Plan plan = (Plan)object;
-
-		return Objects.equals(toString(), plan.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		return PlanSerDes.toJSON(this);
-	}
-
-}
