+++ modules/apps/batch-planner/batch-planner-rest-impl/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Liferay Headless Admin Batch Planner Implementation
-Bundle-SymbolicName: com.liferay.headless.admin.batch.planner.impl
+Bundle-Name: Liferay Batch Planner REST Implementation
+Bundle-SymbolicName: com.liferay.batch.planner.rest.impl
 Bundle-Version: 1.0.0
