+++ modules/apps/batch-planner/batch-planner-rest-impl/build.gradle
@@ -10,7 +10,7 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
 	compileOnly project(":apps:batch-planner:batch-planner-api")
-	compileOnly project(":apps:batch-planner:headless-admin-batch-planner:headless-admin-batch-planner-api")
+	compileOnly project(":apps:batch-planner:batch-planner-rest-api")
 	compileOnly project(":apps:portal-odata:portal-odata-api")
 	compileOnly project(":apps:portal-vulcan:portal-vulcan-api")
 	compileOnly project(":core:petra:petra-function")
