--- modules/apps/batch-planner/batch-planner-rest-impl/src/main/java/com/liferay/headless/admin/batch/planner/internal/jaxrs/application/HeadlessAdminBatchPlannerApplication.java
@@ -1,38 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.admin.batch.planner.internal.jaxrs.application;
-
-import javax.annotation.Generated;
-
-import javax.ws.rs.core.Application;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Matija Petanjek
- * @generated
- */
-@Component(
-	property = {
-		"liferay.jackson=false",
-		"osgi.jaxrs.application.base=/headless-admin-batch-planner",
-		"osgi.jaxrs.extension.select=(osgi.jaxrs.name=Liferay.Vulcan)",
-		"osgi.jaxrs.name=Liferay.Headless.Admin.Batch.Planner"
-	},
-	service = Application.class
-)
-@Generated("")
-public class HeadlessAdminBatchPlannerApplication extends Application {
-}
