--- modules/apps/batch-planner/batch-planner-rest-impl/src/main/java/com/liferay/headless/admin/batch/planner/internal/resource/v1_0/OpenAPIResourceImpl.java
@@ -1,91 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.admin.batch.planner.internal.resource.v1_0;
-
-import com.liferay.portal.vulcan.resource.OpenAPIResource;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.info.Info;
-import io.swagger.v3.oas.annotations.info.License;
-
-import java.util.HashSet;
-import java.util.Set;
-
-import javax.annotation.Generated;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Matija Petanjek
- * @generated
- */
-@Component(
-	properties = "OSGI-INF/liferay/rest/v1_0/openapi.properties",
-	service = OpenAPIResourceImpl.class
-)
-@Generated("")
-@OpenAPIDefinition(
-	info = @Info(license = @License(name = "Apache 2.0", url = "http://www.apache.org/licenses/LICENSE-2.0.html"), title = "Headless Admin Batch Planner", version = "v1.0")
-)
-@Path("/v1.0")
-public class OpenAPIResourceImpl {
-
-	@GET
-	@Path("/openapi.{type:json|yaml}")
-	@Produces({MediaType.APPLICATION_JSON, "application/yaml"})
-	public Response getOpenAPI(@PathParam("type") String type)
-		throws Exception {
-
-		try {
-			Class<? extends OpenAPIResource> clazz =
-				_openAPIResource.getClass();
-
-			clazz.getMethod(
-				"getOpenAPI", Set.class, String.class, UriInfo.class);
-		}
-		catch (NoSuchMethodException noSuchMethodException) {
-			return _openAPIResource.getOpenAPI(_resourceClasses, type);
-		}
-
-		return _openAPIResource.getOpenAPI(_resourceClasses, type, _uriInfo);
-	}
-
-	@Reference
-	private OpenAPIResource _openAPIResource;
-
-	@Context
-	private UriInfo _uriInfo;
-
-	private final Set<Class<?>> _resourceClasses = new HashSet<Class<?>>() {
-		{
-			add(LogResourceImpl.class);
-
-			add(PlanResourceImpl.class);
-
-			add(OpenAPIResourceImpl.class);
-		}
-	};
-
-}
