--- modules/apps/batch-planner/batch-planner-rest-impl/src/main/java/com/liferay/headless/admin/batch/planner/internal/resource/v1_0/PlanResourceImpl.java
@@ -1,173 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.headless.admin.batch.planner.internal.resource.v1_0;
-
-import com.liferay.batch.planner.model.BatchPlannerMapping;
-import com.liferay.batch.planner.model.BatchPlannerPlan;
-import com.liferay.batch.planner.model.BatchPlannerPolicy;
-import com.liferay.batch.planner.service.BatchPlannerMappingService;
-import com.liferay.batch.planner.service.BatchPlannerPlanService;
-import com.liferay.batch.planner.service.BatchPlannerPolicyService;
-import com.liferay.headless.admin.batch.planner.dto.v1_0.Mapping;
-import com.liferay.headless.admin.batch.planner.dto.v1_0.Plan;
-import com.liferay.headless.admin.batch.planner.dto.v1_0.Policy;
-import com.liferay.headless.admin.batch.planner.resource.v1_0.PlanResource;
-import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.pagination.Pagination;
-
-import java.util.List;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ServiceScope;
-
-/**
- * @author Matija Petanjek
- */
-@Component(
-	properties = "OSGI-INF/liferay/rest/v1_0/plan.properties",
-	scope = ServiceScope.PROTOTYPE, service = PlanResource.class
-)
-public class PlanResourceImpl extends BasePlanResourceImpl {
-
-	@Override
-	public void deletePlan(Long id) throws Exception {
-		_batchPlannerPlanService.deleteBatchPlannerPlan(id);
-	}
-
-	@Override
-	public Plan getPlan(Long id) throws Exception {
-		return _toPlan(_batchPlannerPlanService.getBatchPlannerPlan(id));
-	}
-
-	@Override
-	public Page<Plan> getPlansPage(Pagination pagination) throws Exception {
-		return Page.of(
-			transform(
-				_batchPlannerPlanService.getBatchPlannerPlans(
-					contextCompany.getCompanyId(),
-					pagination.getStartPosition(), pagination.getEndPosition()),
-				this::_toPlan),
-			pagination,
-			_batchPlannerPlanService.getBatchPlannerPlansCount(
-				contextCompany.getCompanyId()));
-	}
-
-	@Override
-	public Plan patchPlan(Long id, Plan plan) throws Exception {
-		BatchPlannerPlan batchPlannerPlan =
-			_batchPlannerPlanService.updateBatchPlannerPlan(id, plan.getName());
-
-		Mapping[] mappings = plan.getMappings();
-
-		if (mappings != null) {
-			for (Mapping mapping : plan.getMappings()) {
-				_batchPlannerMappingService.updateBatchPlannerMapping(
-					mapping.getId(), mapping.getExternalFieldName(),
-					mapping.getExternalFieldType(), mapping.getScript());
-			}
-		}
-
-		Policy[] policies = plan.getPolicies();
-
-		if (policies != null) {
-			for (Policy policy : plan.getPolicies()) {
-				_batchPlannerPolicyService.updateBatchPlannerPolicy(
-					id, policy.getName(), policy.getValue());
-			}
-		}
-
-		return _toPlan(batchPlannerPlan);
-	}
-
-	@Override
-	public Plan postPlan(Plan plan) throws Exception {
-		BatchPlannerPlan batchPlannerPlan =
-			_batchPlannerPlanService.addBatchPlannerPlan(
-				plan.getExport(), plan.getExternalType(), plan.getExternalURL(),
-				plan.getInternalClassName(), plan.getName(), false);
-
-		Mapping[] mappings = plan.getMappings();
-
-		if (mappings != null) {
-			for (Mapping mapping : mappings) {
-				_batchPlannerMappingService.addBatchPlannerMapping(
-					batchPlannerPlan.getBatchPlannerPlanId(),
-					mapping.getExternalFieldName(),
-					mapping.getExternalFieldType(),
-					mapping.getInternalFieldName(),
-					mapping.getInternalFieldType(), mapping.getScript());
-			}
-		}
-
-		Policy[] policies = plan.getPolicies();
-
-		if (policies != null) {
-			for (Policy policy : plan.getPolicies()) {
-				_batchPlannerPolicyService.addBatchPlannerPolicy(
-					batchPlannerPlan.getBatchPlannerPlanId(), policy.getName(),
-					policy.getValue());
-			}
-		}
-
-		return _toPlan(batchPlannerPlan);
-	}
-
-	private Mapping[] _getMappings(BatchPlannerPlan batchPlannerPlan)
-		throws Exception {
-
-		List<BatchPlannerMapping> batchPlannerMappings =
-			_batchPlannerMappingService.getBatchPlannerMappings(
-				batchPlannerPlan.getBatchPlannerPlanId());
-
-		return batchPlannerMappings.toArray(new Mapping[0]);
-	}
-
-	private Policy[] _getPolicies(BatchPlannerPlan batchPlannerPlan)
-		throws Exception {
-
-		List<BatchPlannerPolicy> batchPlannerPolicies =
-			_batchPlannerPolicyService.getBatchPlannerPolicies(
-				batchPlannerPlan.getBatchPlannerPlanId());
-
-		return batchPlannerPolicies.toArray(new Policy[0]);
-	}
-
-	private Plan _toPlan(BatchPlannerPlan batchPlannerPlan) throws Exception {
-		return new Plan() {
-			{
-				active = batchPlannerPlan.isActive();
-				export = batchPlannerPlan.isExport();
-				externalType = batchPlannerPlan.getExternalType();
-				externalURL = batchPlannerPlan.getExternalURL();
-				id = batchPlannerPlan.getBatchPlannerPlanId();
-				internalClassName = batchPlannerPlan.getInternalClassName();
-				mappings = _getMappings(batchPlannerPlan);
-				name = batchPlannerPlan.getName();
-				policies = _getPolicies(batchPlannerPlan);
-			}
-		};
-	}
-
-	@Reference
-	private BatchPlannerMappingService _batchPlannerMappingService;
-
-	@Reference
-	private BatchPlannerPlanService _batchPlannerPlanService;
-
-	@Reference
-	private BatchPlannerPolicyService _batchPlannerPolicyService;
-
-}
