+++ modules/apps/batch-planner/batch-planner-rest-test/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Headless Admin Batch Planner Test
-Bundle-SymbolicName: com.liferay.headless.admin.batch.planner.test
+Bundle-Name: Liferay Batch Planner REST Test
+Bundle-SymbolicName: com.liferay.batch.planner.rest.test
 Bundle-Version: 1.0.0
--includeresource: com.liferay.headless.admin.batch.planner.client.jar=com.liferay.headless.admin.batch.planner.client-*.jar;lib:=true
+-includeresource: com.liferay.batch.planner.rest.client.jar=com.liferay.batch.planner.rest.client-*.jar;lib:=true
