+++ modules/apps/batch-planner/batch-planner-rest-test/build.gradle
@@ -3,8 +3,8 @@ dependencies {
 	testIntegrationCompile group: "com.fasterxml.jackson.core", name: "jackson-core", version: "2.12.4"
 	testIntegrationCompile group: "com.fasterxml.jackson.core", name: "jackson-databind", version: "2.12.4"
 	testIntegrationCompile project(":apps:batch-planner:batch-planner-api")
-	testIntegrationCompile project(":apps:batch-planner:headless-admin-batch-planner:headless-admin-batch-planner-api")
-	testIntegrationCompile project(":apps:batch-planner:headless-admin-batch-planner:headless-admin-batch-planner-client")
+	testIntegrationCompile project(":apps:batch-planner:batch-planner-rest-api")
+	testIntegrationCompile project(":apps:batch-planner:batch-planner-rest-client")
 	testIntegrationCompile project(":apps:portal-odata:portal-odata-api")
 	testIntegrationCompile project(":apps:portal-vulcan:portal-vulcan-api")
 	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
